Move all `step` related code into `step.rs`
