+++ ledger/src/proofs/public_input/plonk_checks.rs
@@ -7,8 +7,8 @@ use mina_curves::pasta::Fq;
 use mina_hasher::Fp;
 
 use crate::proofs::{
-    merge::step_verifier::PlonkDomain,
     public_input::plonk_checks::scalars::MinimalForScalar,
+    step::step_verifier::PlonkDomain,
     to_field_elements::ToFieldElements,
     witness::{field, Boolean, FieldWitness, Witness},
     wrap::{wrap_verifier::PlonkWithField, AllFeatureFlags},
